Implementado Geofencing (#38)
* Refector CustomMapFragment variable names

* Make use of Geofencing to draw zones

https://github.com/Fic-Spinoff/family-care/issues/27

It was posible draw circles on map without behaviour. It must be
possible to control if a person enters or exits from one of the drawed
circles.

* Make use of geofencing.
* CustomMapFragment refactores.
* Location and Geofencing as services.

* Make use of Geofencing to draw zones

https://github.com/Fic-Spinoff/family-care/issues/27

It was posible draw circles on map without behaviour. It must be
possible to control if a person enters or exits from one of the drawed
circles.

* Make use of geofencing.
* CustomMapFragment refactores.
* Location and Geofencing as services.

Refactor Geofence service